+++ packages/hub/src/application/sagas/index.ts
@@ -1,7 +1,8 @@
 import { takeLatest } from 'redux-saga/effects';
 import { initApplication } from '../slice/applicationSlice';
 import { initApplicationSaga } from './initApplicationSaga';
+import { manageSagaState } from '../index';
 
 export default function* applicationSaga() {
-  yield takeLatest(initApplication, initApplicationSaga);
+  yield takeLatest(initApplication, manageSagaState(initApplicationSaga));
 }