+++ packages/hub/src/common/arrowLink/ArrowLink.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import cn from 'classnames';
 import { Link, LinkProps } from 'react-router-dom';
-import { ReactComponent as ChevronLeftIcon } from './chevron_left.svg';
-import styles from './LinkButton.module.scss';
+import styles from './ArrowLink.module.scss';
+import { ChevronLeftIcon } from '../icons';
 
 interface LinkButtonProps extends LinkProps {
   hideTextOnMobile?: boolean;
@@ -12,7 +12,7 @@ interface LinkButtonProps extends LinkProps {
   children: React.ReactNode;
 }
 
-const LinkButton: React.FC<LinkButtonProps> = ({
+const ArrowLink: React.FC<LinkButtonProps> = ({
   direction,
   children,
   size = 'medium',
@@ -32,4 +32,4 @@ const LinkButton: React.FC<LinkButtonProps> = ({
   </Link>
 );
 
-export default LinkButton;
+export default ArrowLink;