+++ packages/hub/src/common/divider/Divider.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styles from './Divider.module.scss';
+
+const Divider: React.FC = () => (
+  <div className={styles.divider} />
+);
+
+export default Divider;