+++ packages/hub/src/common/network/slices/networkSlice.ts
@@ -1,0 +1,40 @@
+import { PayloadAction, createSlice } from '@reduxjs/toolkit';
+import isEqual from 'lodash/isEqual';
+import remove from 'lodash/remove';
+
+interface Action { name: string; params?: any; }
+
+export interface NetworkState {
+  actions: Action[],
+  loading: boolean;
+}
+
+const initialState: NetworkState = {
+  actions: [],
+  loading: false,
+};
+
+const networkSlice = createSlice({
+  name: 'network',
+  initialState,
+  reducers: {
+    setLoadingNetworkState: (state: NetworkState, action: PayloadAction<boolean>) => {
+      state.loading = action.payload;
+    },
+    startAction: (state, { payload }: PayloadAction<Action>) => {
+      state.actions.push(payload);
+    },
+    stopAction: (state, { payload: { name, params } }: PayloadAction<Action>) => {
+      remove(state.actions, (a) => a.name === name && isEqual(a.params, params));
+    },
+  },
+});
+
+export const {
+  reducer: networkReducer,
+  actions: {
+    startAction,
+    stopAction,
+    setLoadingNetworkState,
+  },
+} = networkSlice;