+++ packages/hub/src/common/notification/components/Notification.tsx
@@ -2,10 +2,10 @@ import React, { useCallback } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import Snackbar, { SnackbarOrigin } from '@mui/material/Snackbar';
 import Alert from '@mui/material/Alert';
+import { AlertColor } from '@mui/material/Alert/Alert';
 import { hideNotification } from '../slice/notificationSlice';
 import styles from './Notification.module.scss';
 import { ApplicationState } from '../../../application';
-
 
 const mapStateToProps = (state: ApplicationState) => ({
   notification: state.notification.currentNotification,
@@ -45,7 +45,7 @@ const NotificationComponent: React.FC<NotificationProps> = (props) => {
     classes={classes}
     anchorOrigin={options}
   >
-    <Alert variant="filled" onClose={handleClose} severity={notification.type}>
+    <Alert variant="filled" onClose={handleClose} severity={notification.type as AlertColor}>
       {notification.text}
     </Alert>
   </Snackbar>;