+++ packages/hub/src/registration/components/pages/loginRegisterAccount/CreateNewAccount.tsx
@@ -7,8 +7,13 @@ import {
 } from '@mui/material';
 import { SelectChangeEvent } from '@mui/material/Select/SelectInput';
 import classnames from 'classnames';
+import {
+  Tabs,
+  OutlinedInput,
+  checkIfLoading,
+  ModalLoader,
+} from 'common';
 import styles from '../../Registration.module.scss';
-import { Tabs, OutlinedInput } from '../../../../common';
 import {
   setCreatingCurrentShard,
   generateSeedPhrase,
@@ -18,7 +23,10 @@ import {
   setCurrentRegisterCreateAccountTab,
 } from '../../../slice/registrationSlice';
 import {
-  CreateAccountStepsEnum, LoginRegisterAccountTabs, LoginRegisterAccountTabsLabels, RegistrationPageAdditionalProps,
+  CreateAccountStepsEnum,
+  LoginRegisterAccountTabs,
+  LoginRegisterAccountTabsLabels,
+  RegistrationPageAdditionalProps,
 } from '../../../typings/registrationTypes';
 import { getCurrentCreatingStep, getCurrentShardSelector, getGeneratedSeedPhrase } from '../../../selectors/registrationSelectors';
 import { ApplicationState } from '../../../../application';
@@ -30,6 +38,7 @@ const mapStateToProps = (state: ApplicationState) => ({
   currentShard: getCurrentShardSelector(state),
   creatingStep: getCurrentCreatingStep(state),
   generatedSeedPhrase: getGeneratedSeedPhrase(state),
+  loading: checkIfLoading(state, createWallet.type),
 });
 
 const mapDispatchToProps = {
@@ -344,9 +353,13 @@ class CreateNewAccountComponent extends React.PureComponent<CreateNewAccountProp
   };
 
   render() {
-    const { creatingStep } = this.props;
-
+    const { creatingStep, loading } = this.props;
     return <>
+      <ModalLoader
+        loadingTitle={'Processing\nvery soon everything will happen'}
+        open={loading}
+        hideIcon
+      />
       {this.renderContent()}
       <div className={styles.registrationButtonsHolder}>
         {