+++ packages/tssdk/src/evm/examples/loadScNew.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import { EvmApi } from '../../libs';
 import { ChainNameEnum } from '../../config/chain.enum';
 import abiJson from './contractAbi.json';
-import { loadKey } from '../../helpers/key.helper';
+import { CryptoApi } from '../../libs/crypto';
 
 async function testVm() {
   const scAddress = 'AA100000001677721810';
@@ -32,10 +32,10 @@ async function testVm() {
   console.log('getHelloMapArrEl:', getHelloMapArrEl);
 
   // set
-  const key = await loadKey(path.join(__dirname, 'AA100000001677722039.pem'), '1');
+  const key = await CryptoApi.loadKey(path.join(__dirname, 'AA100000001677722039.pem'), '1');
 
   console.log('setGreeting:');
-  await sc.scSet(key, 'setGreeting', ['VALERA']);
+  await sc.scSet(key, 'setGreeting', ['12313']);
 }
 
 testVm()